Update DecisionTree.py
Hey,

A you may know, the code from DecisionTree.py (or the p_decision_tree library, in general) is currently being used in the Introduction to Data Science course in order to illustrate the concepts of decision trees with categorical features. I am a student (Master in Data Science) taking the course and have spend some time exploring the code. In this pull request I propose a change to the visualisation part (from line 222 in the original code, kept unchanged).

The idea is to skip drawing nodes for which the children only have a single child (so skip drawing the parent of leaves and the parent of these parents of leaves) IF all of the leaf nodes share the same value. We aim to achieve that by the shouldDraw()-function which provides a criterion for that. We can then use it to test whether a node is a parent of nodes linked to leaves which are all identical (and consequently draw only the leaf-label). For further details, please look at the example and comments provided.

For further details and discussions you can contact me at: atanas.gruev.98@gmail.com.
From the e-mail address above I send as attachments two pdf-files. The initial one using the code provided to generate a decision tree. The second has the suggested changes, and displays a somewhat-pruned tree which is more straightforward and takes up less space.

Cheers,
Atanas Gruev